Update Flockers example to new format
