+++ examples/Flockers/flockers/boid.py
@@ -1,17 +1,5 @@
-'''
-Flockers
-=============================================================
-A Mesa implementation of Craig Reynolds's Boids flocker model.
-Uses numpy arrays to represent vectors.
-'''
-
-
+from mesa import Agent
 import numpy as np
-import random
-
-from mesa import Model, Agent
-from mesa.space import ContinuousSpace
-from mesa.time import RandomActivation
 
 
 class Boid(Agent):
@@ -99,51 +87,3 @@ class Boid(Agent):
         new_pos = np.array(self.pos) + self.heading * self.speed
         new_x, new_y = new_pos
         model.space.move_agent(self, (new_x, new_y))
-
-
-class BoidModel(Model):
-    '''
-    Flocker model class. Handles agent creation, placement and scheduling.
-    '''
-    N = 100
-    width = 100
-    height = 100
-
-    def __init__(self, N, width, height, speed, vision, separation):
-        '''
-        Create a new Flockers model.
-
-        Args:
-            N: Number of Boids
-            width, height: Size of the space.
-            speed: How fast should the Boids move.
-            vision: How far around should each Boid look for its neighbors
-            separtion: What's the minimum distance each Boid will attempt to
-                       keep from any other
-        '''
-        self.N = N
-        self.vision = vision
-        self.speed = speed
-        self.separation = separation
-        self.schedule = RandomActivation(self)
-        self.space = ContinuousSpace(width, height, True,
-            grid_width=10, grid_height=10)
-        self.make_agents()
-        self.running = True
-
-    def make_agents(self):
-        '''
-        Create N agents, with random positions and starting headings.
-        '''
-        for i in range(self.N):
-            x = random.random() * self.space.x_max
-            y = random.random() * self.space.y_max
-            pos = (x, y)
-            heading = np.random.random(2) * 2 - np.array((1, 1))
-            heading /= np.linalg.norm(heading)
-            boid = Boid(i, pos, self.speed, heading, self.vision, self.separation)
-            self.space.place_agent(boid, pos)
-            self.schedule.add(boid)
-
-    def step(self):
-        self.schedule.step()
