+++ examples/Flockers/flockers/simple_continuous_canvas.js
