+++ examples/Flockers/run.py
@@ -1,0 +1,3 @@
+from flockers.server import server
+
+server.launch()
